Image file upload that is tagged to product posts using Multer(expressjs) as well as pagination logic is added to the web application. Users can now select which page to go to and view the listed product items. Dom Manipulation is also included in this commit, instead of waiting for requests to be done, we simply send data back to the client and have it listen to changes, then simply update/overwrite the current DOM that we see on the frontend

+++ controller/admin.js
@@ -1,3 +1,4 @@
+const fileHelper = require("../util/file");
 const mongodb = require("mongodb");
 const Product = require("../models/product");
 const { validationResult } = require("express-validator/check");
@@ -15,10 +16,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageURL = req.body.imageURL;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add product",
+      path: "/admin/edit-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: "Attached file is not an image",
+      validationErrors: []
+    });
+  }
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -29,7 +44,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageURL: imageURL,
         price: price,
         description: description
       },
@@ -37,6 +51,9 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
+  const imageURL = image.path;
+
   const product = new Product({
     title: title,
     price: price,
@@ -87,7 +104,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImage = req.body.imageURL;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
@@ -100,7 +117,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageURL: updatedImage,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -118,7 +134,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageURL = updatedImage;
+      if (image) {
+        fileHelper.deleteFile(image.path);
+        product.imageURL = image.path;
+      }
       return product.save().then(result => {
         res.redirect("/admin/products");
       });
@@ -128,14 +147,22 @@ exports.postEditProduct = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error("product not found"));
+      }
+      fileHelper(product.imageURL);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
-      res.redirect("/admin/products");
+      res.status(200).json({ message: "Success!" });
     })
     .catch(err => {
-      console.log(err);
+      res.status(500).json({ message: "deleting product failed" });
     });
 };
